% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_pca.R
\name{approx_pca}
\alias{approx_pca}
\title{Approximate the relationship between a genotype and
the principal components of a set of phenotypes.}
\usage{
approx_pca(slopes, stdErrors, intercepts, varX, covPhenos, n, scaled = T,
  meanPhenos = NULL)
}
\arguments{
\item{slopes}{Vector of slopes of linear regressions of each
phenotype as a function of the genotype}

\item{stdErrors}{Vector of standard errors for each slope coefficient}

\item{intercepts}{Vector of intercepts of linear regressions of each
phenotype as a function of the genotype}

\item{varX}{Variance of X, the independent variable (genotype)}

\item{covPhenos}{variance-covariance matrix of all phenotypes}

\item{n}{Sample size}

\item{scaled}{Should summary statistics be converted to correspond to
scaled response variables? (Or, are they already scaled?)}

\item{meanPhenos}{Required for \code{centered = T}. Vector of means
of each phenotype.}

\item{centered}{Center response variables? (Or, already centered?)}
}
\description{
Approximate the relationship between a genotype and
the principal components of a set of phenotypes.
}
