% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_addition.R
\name{covar_adjust}
\alias{covar_adjust}
\title{Adjust for covariates from simple linear regressions.}
\usage{
covar_adjust(coefs, covariateVars, respMean, respVar, n)
}
\arguments{
\item{coefs}{A matrix of coefficients. Each column represents intercept,
and slope coefficient for a linear regression of a common phenotype as a 
function of a single covariates.}

\item{covariateVars}{Variance-covariance matrix of all covariates}

\item{respMean}{Mean of the phenotype}

\item{respVar}{Variance of the phenotype}

\item{n}{Sample size}
}
\description{
Adjust for covariates from simple linear regressions.
}
